Style edit
@johnscancella

Would love to find a way to make the second bullet point on this page a stand-alone sentence with a clearly defined subject. Once it's a full sentence, we can add a period to the end of the bullet (you don't put periods after fragments in a bulleted list), which will look more consistent.

Suggestions welcome, but this is a good page!